Give every catalog sheet one header, and the tweak sheet back its name
The picker sheets and the wrench sheet were two chassis: one measured itself and
carried a title with a close control, the other took a fixed .medium detent, a
nav-bar title and no way out but a swipe. Nesting the first inside the second
made it worse — an outer presentationDetents at the sheet call site overrides the
inner one, so a pushed picker kept its parent's height and left a third of the
sheet empty under Done, beneath a floating circular chevron.

PinwheelSheet is now the one chassis. The leading slot always means get out — a
cross at the root, a back chevron once pushed — the title is centred beside it,
and the sheet measures itself at every depth. Done belongs to a radio group,
which is what a pending selection needs to commit; a list of actions doesn't have
one, so the tweak sheet has no Done and the pickers keep theirs.

The wrench had drifted furthest. It was the component's Options until the display
axes moved into it, which left a per-component control opening a global settings
screen while the same axes sat in the index's bottom bar. It is Tweaks again —
the component's own, named for the vocabulary the code already uses — with the
device button back in the header where it was, and "No tweaks" centred when a
component has none. Theme and appearance stay in the bottom bar.

The header stands 64pt: the 48pt control floor with one spacing-s each side. It
measured 76 before, spacing-l above against spacing-m below, so the title did not
sit centred in its own band.

Tests: unit 123/123 green (local xcodebuild)
